Remove library mode.
All deprecated in v0.31.0 and all unused internally. Removes an entire module
named literally 'library_mode` and a chunk of other confusing code and public
exports we moved away from a long time ago.

Doesn't remove the (already ignored) `--library` cli option.
