+++ docs/manual/src/types/remote_ext_types.md
@@ -54,9 +54,7 @@ enum LogLevel {
 "External types" refer to types defined in other crates that use UniFFI.
 This normally means types from other crates in your workspace.
 
-## Proc-macros
-
-Proc-macro-based code can use external types automatically, without any extra code.
+Proc-macros typically use external types automatically, but UDL needs them described.
 
 ## UDL
 
@@ -125,9 +123,6 @@ which varies slightly for each language:
 
 ### Kotlin
 
-For Kotlin, "library mode" generation with `generate --library [path-to-cdylib]` is recommended when using external types.
-If you use `generate [udl-path]` then the generated code needs to know how to import
-the external types from the Kotlin module that corresponds to the Rust crate.
 By default, UniFFI assumes that the Kotlin module name matches the Rust crate name, but this can be configured in `uniffi.toml` with an entry like this:
 
 ```
